% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulon}
\alias{regulon}
\title{regulon created using \code{epiregulon} package from reprogam-seq data}
\format{
a DFrame.
}
\usage{
data(regulon)
}
\value{
a DFrame.
}
\description{
regulon created using \code{epiregulon} package from reprogam-seq data
}
\examples{
data(regulon)
}
\keyword{datasets}
