% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{enrichPlot}
\alias{enrichPlot}
\title{Plot results of regulonEnrich}
\usage{
enrichPlot(results, top = 15, ncol = 3, title = NULL, combine = TRUE)
}
\arguments{
\item{results}{Output from regulonEnrich}

\item{top}{An integer to indicate the number of pathways to plot ranked by significance. Default is 15.}

\item{ncol}{An integer to indicate the number of columns in the combined plot, if combine == TRUE. Default is 3.}

\item{title}{String indicating the title of the combined plot}

\item{combine}{logical to indicate whether to combine and visualize the plots in one panel}
}
\value{
A combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
Plot results of regulonEnrich
}
\examples{
#retrieve genesets
H <- msigdbr::msigdbr(species="Homo sapiens", collection="H")
genesets <- data.frame(geneset=H$gs_name, genes=H$gene_symbol)
# get regulon
regulon <- data.frame(tf=rep("AR",5),
                      target=c("KLK3","TMPRSS2","KLK2","NKX3-1","FKBP5"),
                      weight=rep(1,5))
enrichment_results <- regulonEnrich(c('AR'), regulon=regulon, genesets=genesets)

# plot graph
enrichPlot(results=enrichment_results )

}
\author{
Xiaosai Yao
}
