% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTests.R
\name{regulonEnrich}
\alias{regulonEnrich}
\title{Perform geneset enrichment of user-defined regulons}
\usage{
regulonEnrich(TF, regulon, weight = "weight", weight_cutoff = 0.5, genesets)
}
\arguments{
\item{TF}{A character vector of TF names}

\item{regulon}{A matrix of weighted regulon consisting of tf, targets, corr and weight}

\item{weight}{String indicating the column name that should be used to filter target genes for geneset enrichment. Default is 'weight'.}

\item{weight_cutoff}{A numeric scalar to indicate the cutoff to filter on the column specified by \code{weight}. Default is 0.5.}

\item{genesets}{A dataframe with the first column being the name of the geneset and the second column being the name of the genes}
}
\value{
A dataframe showing the significantly enriched pathways
}
\description{
Perform geneset enrichment of user-defined regulons
}
\examples{
# retrieve genesets
H <- msigdbr::msigdbr(species="Homo sapiens", collection="H")
genesets <- data.frame(geneset=H$gs_name, genes=H$gene_symbol)
# create some mock regulons
regulon <- data.frame(tf=rep("AR",5),
                      target=c("KLK3","TMPRSS2","KLK2","NKX3-1","FKBP5"),
                      weight=rep(1,5))
enrichment_results <- regulonEnrich(c('AR'), regulon=regulon, genesets=genesets)

}
\author{
Xiaosai Yao
}
